+++ mttl/models/containers/selectors/kv_selector.py
@@ -1,0 +1,192 @@
+import math
+from abc import abstractmethod
+from dataclasses import dataclass
+from typing import Dict
+
+import torch
+from torch import nn
+from torch.nn import functional as F
+
+from mttl.models.containers.selectors.base import Selector, SelectorConfig
+from mttl.models.library.expert import ExpertInfo
+
+
+class KVSelector(Selector):
+    """Selector specific to KV adapters. The KV Adapter modifies the self-attention
+    call, adding the following execution :
+
+    1. adapter_k, adapter_v = adapter.get_kv_weights(k_proj, v_proj)
+    2. adapter_weights = adapter.route(query_states, adapter_k, self)
+        2.1 `adapter.route` calls get_gate(adapter_weights)
+    3. adapter_output = adapter.aggregate(adapter_weights, adapter_v)
+
+    To enable custom routing, one typically needs to modify `get_kv_weights` and `get_gate`.
+    For example, see `KVTaskNameSelector` for an example.
+
+    You can also overwrite the `route` method; the `KVExpertContainer` will call it instead of
+    the default `route` method (see lines 199-201 in `KVExpertContainer`)
+    """
+
+    @abstractmethod
+    def get_kv_weights(self, k_proj, v_proj):
+        pass
+
+    @abstractmethod
+    def get_gate(self, adapter_weights):
+        pass
+
+    def get_merging_weights(self, **selector_kwargs) -> Dict:
+        raise ValueError(
+            f"Not supported for {self.__class__}  since routing depends on input."
+        )
+
+    def on_add_expert(
+        self, expert_name: str, expert_info: ExpertInfo = None, is_default=False
+    ):
+        pass
+
+
+@dataclass
+class KVTaskNameSelectorConfig(SelectorConfig):
+    pass
+
+
+@Selector.register("kv_task_selector", KVTaskNameSelectorConfig)
+class KVTaskNameSelector(KVSelector):
+    """Selects KVAdapters based on the task name."""
+
+    def __init__(self, **kwargs) -> None:
+        super().__init__()
+
+    def get_kv_weights(self, experts, k_proj, v_proj):
+        routing_infos = self.routing_infos
+
+        task_names = routing_infos.task_names
+
+        if task_names is None:
+            task_names = [self.default_expert_name]
+
+        if len(set(task_names)) == 1:
+            # broadcast along batch dim if only 1 task
+            adapter_k, adapter_v = experts[task_names[0]].get_kv_weights(k_proj, v_proj)
+            return adapter_k, adapter_v
+
+        out = zip(
+            *[experts[name].get_kv_weights(k_proj, v_proj) for name in task_names]
+        )
+        out = (torch.cat(tensors, dim=0) for tensors in out)
+        return out
+
+    def get_gate(self, experts, adapter_weights):
+        routing_infos = self.routing_infos
+
+        task_names = routing_infos.task_names
+
+        if task_names is None:
+            task_names = [self.default_expert_name]
+
+        if len(set(task_names)) == 1:
+            # broadcast along batch dim if only 1 task
+            out = experts[task_names[0]].get_gate(adapter_weights)
+            return out
+
+        return torch.cat(
+            [experts[name].get_gate(adapter_weights) for name in task_names],
+        )
+
+
+@dataclass
+class KVConcatSelectorConfig(SelectorConfig):
+    pass
+
+
+@Selector.register("kv_concat_selector", KVConcatSelectorConfig)
+class KVConcatSelector(KVSelector, nn.Module):
+    """Concatenates along the sequence dim. all the adapters, and lets the
+    model's internal attention mechanism take care of routing in a task agnostic way
+    """
+
+    def __init__(self, **kwargs) -> None:
+        super().__init__()
+
+        self.default_expert_name = None
+
+    def get_kv_weights(self, experts, k_proj, v_proj):
+        out = zip(
+            *[
+                kv_adapter.get_kv_weights(k_proj, v_proj)
+                for kv_adapter in experts.values()
+            ]
+        )
+        # NO (1, n_experts, n_heads, soft_prompt_len, head_dim)
+        # (n_experts, n_heads, soft_prompt_len, head_dim)
+        adapter_k, adapter_v = (torch.cat(tensors, dim=0) for tensors in out)
+        n_experts, n_heads, soft_prompt_len, head_dim = adapter_k.size()
+
+        # (n_heads, n_experts * soft_prompt_len, head_dim)
+        adapter_k = adapter_k.transpose(0, 1).reshape(
+            1, n_heads, n_experts * soft_prompt_len, head_dim
+        )
+        adapter_v = adapter_v.transpose(0, 1).reshape(
+            1, n_heads, n_experts * soft_prompt_len, head_dim
+        )
+        return adapter_k, adapter_v
+
+    def get_gate(self, experts, adapter_weights):
+        bsz, n_heads, q_len, n_exp_kv_len = adapter_weights.size()
+        adapter_weights = adapter_weights.view(bsz, n_heads, q_len, len(experts), -1)
+
+        # sum probs over all `soft_prompt_len` keys, to get (bsz, n_heads, q_len, n_experts)
+        per_expert_weight = adapter_weights.sum(dim=-1)
+
+        # (n_experts, n_heads, 1, 1)
+        all_gates = torch.cat(
+            [kv_adapter.get_gate(adapter_weights) for kv_adapter in experts.values()]
+        )
+        # output : (bsz, n_heads, q_len, 1)
+        out = torch.einsum("bhqe,ehab->bhqa", per_expert_weight, all_gates)
+        return out
+
+
+@dataclass
+class KVNormSelectorConfig(SelectorConfig):
+    pass
+
+
+@Selector.register("kv_norm_selector", KVNormSelectorConfig)
+class KVNormSelector(KVSelector):
+    def route(self, experts, query, keys, attn_layer):
+        """(2) Compute The Standard Attention Scores in augmented attention"""
+
+        query = F.normalize(query, dim=-1, p=2)
+        keys = F.normalize(keys, dim=-1, p=2)
+
+        adapter_logits = torch.matmul(
+            query, keys.transpose(2, 3).type_as(query)
+        ) / math.sqrt(attn_layer.head_dim)
+
+        adapter_weights = F.softmax(adapter_logits, dim=-1, dtype=torch.float32)
+        gate_out = self.get_gate(experts, adapter_weights)
+        out = gate_out * adapter_weights.type_as(query)
+
+        return out
+
+
+@dataclass
+class KVConcatNormSelectorConfig(SelectorConfig):
+    pass
+
+
+@Selector.register("kv_concat_norm_selector", KVConcatNormSelectorConfig)
+class KVConcatNormSelector(KVConcatSelector, KVNormSelector):
+    pass
+
+
+@dataclass
+class KVTaskNameNormSelectorConfig(SelectorConfig):
+    pass
+
+
+@Selector.register("kv_task_norm_selector", KVTaskNameNormSelectorConfig)
+class KVTaskNameNormSelector(KVTaskNameSelector, KVNormSelector):
+    pass
